+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Features
+- You can now pass options `-t/--tty` and `-i/--interactive` to `armada ssh`, which translate to corresponding
+`docker exec` options. Also the default behavior has been changed in case the command is provided. Both options,
+`-t` and `-i`, are turned off then.
+
 ### Bug fixes
 - Fix bug with `armada ssh` always returning exit code 0.
 - Fix recovering services after Docker/host restart.